requirements.txt
streamlit==1.34.0
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
langchain==0.2.5
langchain-community==0.2.5
langchain-google-genai==1.0.5
pypdf==4.2.0
sentence-transformers==2.7.0
google-generativeai==0.5.4
faiss-cpu==1.7.4
pymupdf==1.24.5
pillow==10.3.0
python-dotenv==1.0.1
requests==2.31.0